+++ src/jsifier.js
@@ -17,6 +17,9 @@ global.proxiedFunctionTable = ['null'/* Reserve index 0 for an undefined functio
 
 // Mangles the given C/JS side function name to assembly level function name (adds an underscore)
 function mangleCSymbolName(f) {
+  if (f === '__main_argc_argv') {
+    f = 'main';
+  }
   return f[0] == '$' ? f.substr(1) : '_' + f;
 }
 
@@ -51,6 +54,9 @@ function stringifyWithFunctions(obj) {
 
 function isDefined(symName) {
   if (WASM_EXPORTS.has(symName) || SIDE_MODULE_EXPORTS.has(symName)) {
+    return true;
+  }
+  if (symName == '__main_argc_argv' && SIDE_MODULE_EXPORTS.has('main')) {
     return true;
   }
   // 'invoke_' symbols are created at runtime in libary_dylink.py so can
@@ -197,16 +203,24 @@ function ${name}(${args}) {
         return '';
       }
 
-      let noExport = false;
+      // This gets set to true in the case of dynamic linking for symbols that
+      // are undefined in the main module.  In this case we create a stub that
+      // will resolve the correct symbol at runtime, or assert if its missing.
+      let isStub = false;
 
       if (!LibraryManager.library.hasOwnProperty(ident)) {
         if (ONLY_CALC_JS_SYMBOLS) {
           return;
         }
-        if (!isDefined(ident)) {
+        const isWeakImport = WEAK_IMPORTS.has(ident);
+        if (!isDefined(ident) && !isWeakImport) {
+          if (PROXY_TO_PTHREAD && !MAIN_MODULE && ident == '__main_argc_argv') {
+            error('PROXY_TO_PTHREAD proxies main() for you, but no main exists');
+            return;
+          }
           let undefinedSym = ident;
           if (ident === '__main_argc_argv') {
-            undefinedSym = 'main';
+            undefinedSym = 'main/__main_argc_argv';
           }
           let msg = 'undefined symbol: ' + undefinedSym;
           if (dependent) msg += ` (referenced by ${dependent})`;
@@ -220,7 +234,7 @@ function ${name}(${args}) {
           } else if (VERBOSE || WARN_ON_UNDEFINED_SYMBOLS) {
             warn(msg);
           }
-          if (undefinedSym === 'main' && STANDALONE_WASM) {
+          if (ident === '__main_argc_argv' && STANDALONE_WASM) {
             warn('To build in STANDALONE_WASM mode without a main(), use emcc --no-entry');
           }
         }
@@ -239,7 +253,7 @@ function ${name}(${args}) {
           }
           const functionBody = assertion + `return ${target}.apply(null, arguments);`;
           LibraryManager.library[ident] = new Function(functionBody);
-          noExport = true;
+          isStub = true;
         }
       }
 
@@ -383,7 +397,7 @@ function ${name}(${args}) {
       const sig = LibraryManager.library[ident + '__sig'];
       // asm module exports are done in emscripten.py, after the asm module is ready. Here
       // we also export library methods as necessary.
-      if ((EXPORT_ALL || EXPORTED_FUNCTIONS.has(finalName)) && !noExport) {
+      if ((EXPORT_ALL || EXPORTED_FUNCTIONS.has(finalName)) && !isStub) {
         contentText += `\nModule["${finalName}"] = ${finalName};`;
       }
       // Relocatable code needs signatures to create proper wrappers. Stack
@@ -393,6 +407,9 @@ function ${name}(${args}) {
       //       of async imports/exports.
       if (sig && (RELOCATABLE || ASYNCIFY == 2)) {
         contentText += `\n${finalName}.sig = '${sig}';`;
+      }
+      if (isStub) {
+        contentText += `\n${finalName}.stub = true;`;
       }
 
       let commentText = '';